backend/requirements.txt
django==5.1.5
django-cors-headers==4.6.0
django-utils-six==2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
python-dotenv==1.0.1
environs==14.1.0
requests==2.32.3
dj-database-url==2.3.0
langchain-google-genai==2.0.9
beautifulsoup4==4.13.3
